Refactor NumberOfOccurences
