+++ src/main/java/seedu/address/model/lesson/Lesson.java
@@ -1,5 +1,7 @@
 package seedu.address.model.lesson;
 
+import seedu.address.model.components.occurences.NumberOfOccurrences;
+
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 
 import java.time.LocalTime;
