+++ src/test/java/seedu/address/model/lesson/NumberOfOccurrencesTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
+import seedu.address.model.components.occurences.NumberOfOccurrences;
 
 public class NumberOfOccurrencesTest {
